# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Running Academy API
  description: |
    Comprehensive API for the Running Academy fitness tracking application.

    ## Features
    - User authentication and profile management
    - Real-time workout tracking with GPS
    - Social features and leaderboards
    - Personalized coaching tips
    - Device integration and notifications

    ## Authentication
    Most endpoints require JWT Bearer token authentication.

    ## Base URL
    - Development: `http://localhost:3001`
    - Production: `https://api.runningacademy.com`
  version: 1.0.0
  contact:
    name: Running Academy Team
    email: api@runningacademy.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001
    description: Development server
  - url: https://api.runningacademy.com
    description: Production server
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          description: Unique user identifier
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        birthDate:
          type: string
          format: date
        gender:
          type: string
          enum: [male, female, other]
        weight:
          type: number
          description: Weight in kg
        height:
          type: number
          description: Height in cm
        fitnessLevel:
          type: string
          enum: [beginner, intermediate, advanced]
        weeklyGoal:
          type: integer
          description: Weekly distance goal in km
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
        timestamp:
          type: string
          format: date-time
paths:
...
  /api/auth/register:
    post:
      tags: [Authentication]
      summary: Register new user
      description: Create a new user account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                [firstName, lastName, email, password, birthDate, gender]
              properties:
                firstName:
                  type: string
                  example: John
                lastName:
                  type: string
                  example: Runner
                email:
                  type: string
                  format: email
                  example: john@example.com
                password:
                  type: string
                  minLength: 8
                  example: password123
                birthDate:
                  type: string
                  format: date
                  example: 1990-01-01
                gender:
                  type: string
                  enum: [male, female, other]
                  example: male
                weight:
                  type: number
                  example: 70
                height:
                  type: number
                  example: 175
                fitnessLevel:
                  type: string
                  enum: [beginner, intermediate, advanced]
                  example: intermediate
      responses:
        "201":
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Registration successful
                  token:
                    type: string
                  user:
                    $ref: "#/components/schemas/User"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: User already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/auth/login:
    post:
      tags: [Authentication]
      summary: User login
      description: Authenticate user and return JWT token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                  example: john@example.com
                password:
                  type: string
                  example: password123
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  token:
                    type: string
                  user:
                    $ref: "#/components/schemas/User"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...